+++ NEWS.md
@@ -4,7 +4,7 @@
 
 - `knitr::knit()` starts to warn against non-UTF8 encodings ([Why?](https://yihui.name/en/2018/11/biggest-regret-knitr/)). In the future, we will only support UTF-8 input. If your input file is not encoded in UTF-8, we strongly recommend that you re-save it with UTF-8.
 
-- Removed the `encoding` arguments in `knitr::pandoc()` and `knitr::Sweave2knitr()`. The input files must be encoded in UTF-8.
+- Removed the `encoding` arguments in `knitr::pandoc()`, `knitr::knit2pandoc()`, `knitr::knit2pdf()`, and `knitr::Sweave2knitr()`. The input files must be encoded in UTF-8.
 
 # CHANGES IN knitr VERSION 1.21